# Conflicts:
#	src/AElf.OS.Core/Network/Infrastructure/IPeer.cs
#	src/AElf.OS.Core/Network/NetworkConstants.cs
#	src/AElf.OS.Network.Grpc/Connection/ConnectionInfoProvider.cs
#	src/AElf.OS.Network.Grpc/GrpcNetworkModule.cs
#	src/AElf.OS.Network.Grpc/GrpcPeer.cs
#	test/AElf.OS.Network.Grpc.Tests/GrpcBasicNetworkTestModule.cs
#	test/AElf.OS.Network.Grpc.Tests/NetworkTestContextHelpers.cs

+++ src/AElf.OS.Network.Grpc/GrpcPeer.cs
@@ -6,6 +6,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
+using System.Threading.Tasks.Dataflow;
 using AElf.Kernel;
 using AElf.OS.Network.Application;
 using AElf.OS.Network.Infrastructure;
@@ -23,12 +24,15 @@ namespace AElf.OS.Network.Grpc
     public class GrpcPeer : IPeer
     {
         private const int MaxMetricsPerMethod = 100;
-
-        private const int BlockRequestTimeout = 300;
-        private const int BlocksRequestTimeout = 500;
+        private const int BlockRequestTimeout = 500;
+        private const int BlocksRequestTimeout = 800;
         private const int GetNodesTimeout = 500;
-
         private const int UpdateHandshakeTimeout = 400;
+        private const int StreamRecoveryWaitTimeInMilliseconds = 500;
+
+        private const int MaxDegreeOfParallelismForAnnouncementJobs = 3;
+        private const int MaxDegreeOfParallelismForTransactionJobs = 1;
+        private const int MaxDegreeOfParallelismForBlockJobs = 1;
 
         private enum MetricNames
         {
@@ -41,7 +45,8 @@ namespace AElf.OS.Network.Grpc
         private readonly PeerService.PeerServiceClient _client;
 
         /// <summary>
-        /// Property that describes a valid state. Valid here means that the peer is ready to be used for communications.
+        /// Property that describes that describes if the peer is ready for send/request operations. It's based
+        /// on the state of the underlying channel and the IsConnected.
         /// </summary>
         public bool IsReady
         {
@@ -67,6 +72,11 @@ namespace AElf.OS.Network.Grpc
         public long CurrentBlockHeight { get; private set; }
 
         public IPEndPoint RemoteEndpoint { get; }
+        public int BufferedTransactionsCount => _sendTransactionJobs.InputCount;
+        public int BufferedBlocksCount => _sendBlockJobs.InputCount;
+        public int BufferedAnnouncementsCount => _sendAnnouncementJobs.InputCount;
+
+        public string IpAddress { get; }
 
         public PeerInfo Info { get; }
 
@@ -79,6 +89,10 @@ namespace AElf.OS.Network.Grpc
         private AsyncClientStreamingCall<Transaction, VoidReply> _transactionStreamCall;
         private AsyncClientStreamingCall<BlockAnnouncement, VoidReply> _announcementStreamCall;
         private AsyncClientStreamingCall<BlockWithTransactions, VoidReply> _blockStreamCall;
+
+        private readonly ActionBlock<StreamJob> _sendAnnouncementJobs;
+        private readonly ActionBlock<StreamJob> _sendBlockJobs;
+        private readonly ActionBlock<StreamJob> _sendTransactionJobs;
 
         public GrpcPeer(GrpcClient client, IPEndPoint remoteEndpoint, PeerInfo peerInfo)
         {
@@ -98,12 +112,30 @@ namespace AElf.OS.Network.Grpc
             _recentRequestsRoundtripTimes.TryAdd(nameof(MetricNames.Announce), new ConcurrentQueue<RequestMetric>());
             _recentRequestsRoundtripTimes.TryAdd(nameof(MetricNames.GetBlock), new ConcurrentQueue<RequestMetric>());
             _recentRequestsRoundtripTimes.TryAdd(nameof(MetricNames.GetBlocks), new ConcurrentQueue<RequestMetric>());
+
+            _sendAnnouncementJobs = new ActionBlock<StreamJob>(SendStreamJobAsync,
+                new ExecutionDataflowBlockOptions
+                {
+                    MaxDegreeOfParallelism = MaxDegreeOfParallelismForAnnouncementJobs,
+                    BoundedCapacity = NetworkConstants.DefaultMaxBufferedAnnouncementCount
+                });
+            _sendBlockJobs = new ActionBlock<StreamJob>(SendStreamJobAsync,
+                new ExecutionDataflowBlockOptions
+                {
+                    MaxDegreeOfParallelism = MaxDegreeOfParallelismForBlockJobs,
+                    BoundedCapacity = NetworkConstants.DefaultMaxBufferedBlockCount
+                });
+            _sendTransactionJobs = new ActionBlock<StreamJob>(SendStreamJobAsync,
+                new ExecutionDataflowBlockOptions
+                {
+                    MaxDegreeOfParallelism = MaxDegreeOfParallelismForTransactionJobs,
+                    BoundedCapacity = NetworkConstants.DefaultMaxBufferedTransactionCount
+                });
         }
 
         public Dictionary<string, List<RequestMetric>> GetRequestMetrics()
         {
             Dictionary<string, List<RequestMetric>> metrics = new Dictionary<string, List<RequestMetric>>();
-
             foreach (var roundtripTime in _recentRequestsRoundtripTimes.ToArray())
             {
                 var metricsToAdd = new List<RequestMetric>();
@@ -123,7 +155,7 @@ namespace AElf.OS.Network.Grpc
             GrpcRequest request = new GrpcRequest {ErrorMessage = "Request nodes failed."};
             Metadata data = new Metadata {{GrpcConstants.TimeoutMetadataKey, GetNodesTimeout.ToString()}};
 
-            return RequestAsync(_client, c => c.GetNodesAsync(new NodesRequest {MaxCount = count}, data), request);
+            return RequestAsync(() => _client.GetNodesAsync(new NodesRequest {MaxCount = count}, data), request);
         }
 
         public void UpdateLastReceivedHandshake(Handshake handshake)
@@ -146,7 +178,7 @@ namespace AElf.OS.Network.Grpc
 
             Metadata data = new Metadata {{GrpcConstants.TimeoutMetadataKey, BlockRequestTimeout.ToString()}};
 
-            var blockReply = await RequestAsync(_client, c => c.RequestBlockAsync(blockRequest, data), request);
+            var blockReply = await RequestAsync(() => _client.RequestBlockAsync(blockRequest, data), request);
 
             return blockReply?.Block;
         }
@@ -165,7 +197,7 @@ namespace AElf.OS.Network.Grpc
 
             Metadata data = new Metadata {{GrpcConstants.TimeoutMetadataKey, BlocksRequestTimeout.ToString()}};
 
-            var list = await RequestAsync(_client, c => c.RequestBlocksAsync(blockRequest, data), request);
+            var list = await RequestAsync(() => _client.RequestBlocksAsync(blockRequest, data), request);
 
             if (list == null)
                 return new List<BlockWithTransactions>();
@@ -175,11 +207,65 @@ namespace AElf.OS.Network.Grpc
 
         #region Streaming
 
-        public async Task SendBlockAsync(BlockWithTransactions blockWithTransactions)
+        public void EnqueueTransaction(Transaction transaction, Action<NetworkException> sendCallback)
         {
-            if (!IsConnected)
+            if (!IsReady)
+                throw new NetworkException($"Dropping transaction, peer is not ready - {this}.",
+                    NetworkExceptionType.NotConnected);
+
+            _sendTransactionJobs.Post(new StreamJob{Transaction = transaction, SendCallback = sendCallback});
+        }
+
+        public void EnqueueAnnouncement(BlockAnnouncement announcement, Action<NetworkException> sendCallback)
+        {
+            if (!IsReady)
+                throw new NetworkException($"Dropping announcement, peer is not ready - {this}.",
+                    NetworkExceptionType.NotConnected);
+
+            _sendAnnouncementJobs.Post(new StreamJob {BlockAnnouncement = announcement, SendCallback = sendCallback});
+        }
+
+        public void EnqueueBlock(BlockWithTransactions blockWithTransactions, Action<NetworkException> sendCallback)
+        {
+            if (!IsReady)
+                throw new NetworkException($"Dropping block, peer is not ready - {this}.",
+                    NetworkExceptionType.NotConnected);
+
+            _sendBlockJobs.Post(new StreamJob{BlockWithTransactions = blockWithTransactions, SendCallback = sendCallback});
+        }
+
+        private async Task SendStreamJobAsync(StreamJob job)
+        {
+            if (!IsReady)
                 return;
 
+            try
+            {
+                if (job.Transaction != null)
+                {
+                    await SendTransactionAsync(job.Transaction);
+                }
+                else if (job.BlockAnnouncement != null)
+                {
+                    await SendAnnouncementAsync(job.BlockAnnouncement);
+                }
+                else if (job.BlockWithTransactions != null)
+                {
+                    await BroadcastBlockAsync(job.BlockWithTransactions);
+                }
+            }
+            catch (RpcException ex)
+            {
+                job.SendCallback?.Invoke(CreateNetworkException(ex, $"Error on broadcast to {this}: "));
+                await Task.Delay(StreamRecoveryWaitTimeInMilliseconds);
+                return;
+            }
+
+            job.SendCallback?.Invoke(null);
+        }
+
+        private async Task BroadcastBlockAsync(BlockWithTransactions blockWithTransactions)
+        {
             if (_blockStreamCall == null)
                 _blockStreamCall = _client.BlockBroadcastStream();
 
@@ -187,12 +273,12 @@ namespace AElf.OS.Network.Grpc
             {
                 await _blockStreamCall.RequestStream.WriteAsync(blockWithTransactions);
             }
-            catch (RpcException e)
+            catch (RpcException)
             {
                 _blockStreamCall.Dispose();
                 _blockStreamCall = null;
 
-                HandleFailure(e, $"Error during block broadcast: {blockWithTransactions.Header.GetHash()}.");
+                throw;
             }
         }
 
@@ -200,11 +286,8 @@ namespace AElf.OS.Network.Grpc
         /// Send a announcement to the peer using the stream call.
         /// Note: this method is not thread safe.
         /// </summary>
-        public async Task SendAnnouncementAsync(BlockAnnouncement header)
+        private async Task SendAnnouncementAsync(BlockAnnouncement header)
         {
-            if (!IsConnected)
-                return;
-
             if (_announcementStreamCall == null)
                 _announcementStreamCall = _client.AnnouncementBroadcastStream();
 
@@ -212,25 +295,21 @@ namespace AElf.OS.Network.Grpc
             {
                 await _announcementStreamCall.RequestStream.WriteAsync(header);
             }
-            catch (RpcException e)
+            catch (RpcException)
             {
                 _announcementStreamCall.Dispose();
                 _announcementStreamCall = null;
 
-                HandleFailure(e, $"Error during announcement broadcast: {header.BlockHash}.");
+                throw;
             }
         }
-
 
         /// <summary>
         /// Send a transaction to the peer using the stream call.
         /// Note: this method is not thread safe.
         /// </summary>
-        public async Task SendTransactionAsync(Transaction transaction)
+        private async Task SendTransactionAsync(Transaction transaction)
         {
-            if (!IsConnected)
-                return;
-
             if (_transactionStreamCall == null)
                 _transactionStreamCall = _client.TransactionBroadcastStream();
 
@@ -238,17 +317,17 @@ namespace AElf.OS.Network.Grpc
             {
                 await _transactionStreamCall.RequestStream.WriteAsync(transaction);
             }
-            catch (RpcException e)
+            catch (RpcException)
             {
                 _transactionStreamCall.Dispose();
                 _transactionStreamCall = null;
 
-                HandleFailure(e, $"Error during transaction broadcast: {transaction.GetHash()}.");
+                throw;
             }
         }
 
         #endregion
-
+        
         public async Task ConfirmHandshakeAsync()
         {
             var request = new GrpcRequest {ErrorMessage = "Error while sending confirm handshake."};
@@ -258,15 +337,13 @@ namespace AElf.OS.Network.Grpc
                 {GrpcConstants.TimeoutMetadataKey, UpdateHandshakeTimeout.ToString()}
             };
 
-            await RequestAsync(_client, c => c.ConfirmHandshakeAsync(new ConfirmHandshakeRequest(), data), request);
+            await RequestAsync(() => _client.ConfirmHandshakeAsync(new ConfirmHandshakeRequest(), data), request);
         }
 
-        // todo consider removing client from the lambda as it is not used. It can be capture by the func.
-        private async Task<TResp> RequestAsync<TResp>(PeerService.PeerServiceClient client,
-            Func<PeerService.PeerServiceClient, AsyncUnaryCall<TResp>> func, GrpcRequest requestParams)
+        private async Task<TResp> RequestAsync<TResp>(Func<AsyncUnaryCall<TResp>> func, GrpcRequest requestParams)
         {
             var metricsName = requestParams.MetricName;
-            bool timeRequest = !string.IsNullOrEmpty(metricsName);
+            var timeRequest = !string.IsNullOrEmpty(metricsName);
             var requestStartTime = TimestampHelper.GetUtcNow();
 
             Stopwatch requestTimer = null;
@@ -276,8 +353,7 @@ namespace AElf.OS.Network.Grpc
 
             try
             {
-                var response = await func(client);
-
+                var response = await func();
                 if (timeRequest)
                 {
                     requestTimer.Stop();
@@ -286,9 +362,9 @@ namespace AElf.OS.Network.Grpc
 
                 return response;
             }
-            catch (AggregateException e)
+            catch (AggregateException ex)
             {
-                HandleFailure(e.Flatten(), requestParams.ErrorMessage);
+                throw CreateNetworkException(ex.Flatten(), requestParams.ErrorMessage);
             }
             finally
             {
@@ -298,8 +374,6 @@ namespace AElf.OS.Network.Grpc
                     RecordMetric(requestParams, requestStartTime, requestTimer.ElapsedMilliseconds);
                 }
             }
-
-            return default(TResp);
         }
 
         private void RecordMetric(GrpcRequest grpcRequest, Timestamp requestStartTime, long elapsedMilliseconds)
@@ -322,39 +396,59 @@ namespace AElf.OS.Network.Grpc
         /// This method handles the case where the peer is potentially down. If the Rpc call
         /// put the channel in TransientFailure or Connecting, we give the connection a certain time to recover.
         /// </summary>
-        private void HandleFailure(Exception exception, string errorMessage)
+        private NetworkException CreateNetworkException(Exception exception, string errorMessage)
         {
-            // If channel has been shutdown (unrecoverable state) remove it.
             string message = $"Failed request to {this}: {errorMessage}";
             NetworkExceptionType type = NetworkExceptionType.Rpc;
 
-            if (_channel.State == ChannelState.Shutdown)
+            if (_channel.State != ChannelState.Ready)
             {
-                message = $"Peer is shutdown - {this}: {errorMessage}";
-                type = NetworkExceptionType.Unrecoverable;
+                // if channel has been shutdown (unrecoverable state) remove it.
+                if (_channel.State == ChannelState.Shutdown)
+                {
+                    message = $"Peer is shutdown - {this}: {errorMessage}";
+                    type = NetworkExceptionType.Unrecoverable;
+                }
+                else if (_channel.State == ChannelState.TransientFailure || _channel.State == ChannelState.Connecting)
+                {
+                    // from this we try to recover
+                    message = $"Peer is unstable - {this}: {errorMessage}";
+                    type = NetworkExceptionType.PeerUnstable;
+                }
+                else
+                {
+                    // if idle just after an exception, disconnect.
+                    message = $"Peer error, channel state {_channel.State} - {this}: {errorMessage}";
+                    type = NetworkExceptionType.Unrecoverable;
+                }
             }
-            else if (_channel.State == ChannelState.TransientFailure || _channel.State == ChannelState.Connecting)
+            else
             {
-                message = $"Failed request to {this}: {errorMessage}";
-                type = NetworkExceptionType.PeerUnstable;
-            }
-            else if (exception.InnerException is RpcException rpcEx && rpcEx.StatusCode == StatusCode.Cancelled)
-            {
-                message = $"Failed request to {this}: {errorMessage}";
-                type = NetworkExceptionType.Unrecoverable;
+                // there was an exception, not related to connectivity.
+                if (exception.InnerException is RpcException rpcEx && rpcEx.StatusCode == StatusCode.Cancelled)
+                {
+                    message = $"Request was cancelled {this}: {errorMessage}";
+                    type = NetworkExceptionType.Unrecoverable;
+                }
             }
 
-            throw new NetworkException(message, exception, type);
+            return new NetworkException(message, exception, type);
         }
 
         public async Task<bool> TryRecoverAsync()
         {
+            if (_channel.State == ChannelState.Shutdown)
+                return false;
+
             await _channel.TryWaitForStateChangedAsync(_channel.State,
-                DateTime.UtcNow.AddSeconds(NetworkConstants.DefaultPeerDialTimeoutInMilliSeconds));
+                DateTime.UtcNow.AddSeconds(NetworkConstants.DefaultPeerRecoveryTimeoutInMilliSeconds));
 
             // Either we connected again or the state change wait timed out.
             if (_channel.State == ChannelState.TransientFailure || _channel.State == ChannelState.Connecting)
+            {
+                IsConnected = false;
                 return false;
+            }
 
             return true;
         }
@@ -381,17 +475,26 @@ namespace AElf.OS.Network.Grpc
             IsConnected = false;
             IsShutdown = true;
 
+            // we complete but no need to await the jobs
+            _sendAnnouncementJobs.Complete();
+            _sendBlockJobs.Complete();
+            _sendTransactionJobs.Complete();
+
+            _announcementStreamCall?.Dispose();
+            _transactionStreamCall?.Dispose();
+            _blockStreamCall?.Dispose();
+
             // send disconnect message if the peer is still connected and the connection
             // is stable.
-            if (gracefulDisconnect && IsReady)
+            if (gracefulDisconnect && (_channel.State == ChannelState.Idle || _channel.State == ChannelState.Ready))
             {
                 GrpcRequest request = new GrpcRequest {ErrorMessage = "Error while sending disconnect."};
 
                 try
                 {
-                    await RequestAsync(_client,
-                        c => c.DisconnectAsync(new DisconnectReason {Why = DisconnectReason.Types.Reason.Shutdown}),
-                        request);
+                    await RequestAsync(
+                        () => _client.DisconnectAsync(new DisconnectReason
+                            {Why = DisconnectReason.Types.Reason.Shutdown}), request);
                 }
                 catch (NetworkException)
                 {
